Fix error while locating features directory
When cloning from repository, _template is set to repo URL in cookiecutter context. Thus, it cannot be used to locate local repo direcroty.

The only way so far is to use the directory from cookiecutter config. But to do it reliably, we need to force default config. Otherwise, user config could break this functionality.